Coverage: Introduce election/candidacy meta data
Given we may have multiple elections and candidacies at any one time, and press coverage could relate to individual elections/candidacies this is a proposal to add appropriate meta data to coverage items.

If this is acceptable/workable/not a ridiculous idea, my hope would be to use it to ingest election-specific coverage in other places. For example, the [People Power Brum website](https://peoplepowerbrum.com) could implement a dynamically-updating list of coverage on it's WordPress site.