Update code_of_conduct.md
Suggest removing "political attacks" from the examples of unacceptable behaviour as it is ambiguous.  Political attacks in the sense of dispassionate attack of political views should be acceptable as part of healthy debate.  By contrast, the use of politics as an excuse or pretext for a personal attack is entirely unacceptable, but the other examples already make that clear.